Move entities to TimeScout.Domain project
