+++ TimeScout.Domain/Entities/Event.cs
@@ -1,6 +1,4 @@
-using TimeScout.API.Entity;
-
-namespace TimeScout.API.Models;
+namespace TimeScout.Domain.Entities;
 
 public class Event : BaseEntity
 {
